+++ src/main/java/com/github/mnesikos/simplycats/configuration/SCConfig.java
@@ -11,7 +11,9 @@ public class SCConfig {
     private static final String PREFIX = "config." + SimplyCats.MOD_ID;
     public static ForgeConfigSpec.BooleanValue join_message;
     public static ForgeConfigSpec.BooleanValue attack_ai;
-    public static ForgeConfigSpec.ConfigValue<List<? extends String>> prey_list;
+    public static ForgeConfigSpec.BooleanValue replace_tamed_vanilla;
+    public static ForgeConfigSpec.BooleanValue replace_vanilla_spawns;
+    public static ForgeConfigSpec.BooleanValue intact_stray_spawns;
     public static ForgeConfigSpec.ConfigValue<Double> wander_area_limit;
     public static ForgeConfigSpec.ConfigValue<Integer> tamed_limit;
     public static ForgeConfigSpec.ConfigValue<Integer> breeding_limit;
@@ -20,6 +22,7 @@ public class SCConfig {
     public static ForgeConfigSpec.ConfigValue<Integer> heat_timer;
     public static ForgeConfigSpec.ConfigValue<Integer> heat_cooldown;
     public static ForgeConfigSpec.ConfigValue<Integer> male_cooldown;
+    public static ForgeConfigSpec.ConfigValue<List<? extends String>> prey_list;
 
     static {
         ForgeConfigSpec.Builder configBuilder = new ForgeConfigSpec.Builder();
@@ -29,6 +32,21 @@ public class SCConfig {
 
     private static void setupConfig(ForgeConfigSpec.Builder builder) {
         builder.push("Options");
+        replace_tamed_vanilla = builder.worldRestart()
+                .comment(" If you have existing tamed vanilla cats in your world and want to change them to Simply Cats, enable this and restart your minecraft instance.")
+                .translation(PREFIX + ".replace_tamed_vanilla")
+                .define("replace_tamed_vanilla", false);
+
+        replace_vanilla_spawns = builder
+                .comment(" Enable or disable replacing vanilla cat spawns with Simply Cats.")
+                .translation(PREFIX + ".replace_vanilla_spawns")
+                .define("replace_vanilla_spawns", true);
+
+        intact_stray_spawns = builder
+                .comment(" Disable this if you want stray village cats spawning fixed (enabled for intact cat spawns).")
+                .translation(PREFIX + ".intact_stray_spawns")
+                .define("intact_stray_spawns", true);
+
         join_message = builder
                 .comment(" Enable or disable the initial join message with a player's cat count.")
                 .translation(PREFIX + ".join_message")
@@ -37,10 +55,11 @@ public class SCConfig {
         attack_ai = builder
                 .comment(" Disabling this will not allow cats to attack entities in their prey list, essentially a peaceful mode for cats.")
                 .translation(PREFIX + ".attack_ai")
-                .define("ATTACK_AI", true);
+                .define("attack_ai", true);
 
         wander_area_limit = builder.worldRestart()
-                .comment(" When a cat's home is set, this is the distance in blocks they are allowed to roam.")
+                .comment(" When a cat's home is set, this is the distance in blocks they are allowed to roam.",
+                        " Default: 400.0")
                 .translation(PREFIX + ".wander_area_limit")
                 .define("wander_area_limit", 400.0D);
 
@@ -50,7 +69,8 @@ public class SCConfig {
                 .define("tamed_limit", 0);
 
         breeding_limit = builder
-                .comment(" This number is used to limit cat breeding; if more than this amount of cats are nearby, automatic breeding will be disabled.")
+                .comment(" This number is used to limit cat breeding; if more than this amount of cats are nearby, automatic breeding will be disabled.",
+                        " Default: 20")
                 .translation(PREFIX + ".breeding_limit")
                 .define("breeding_limit", 20);
         builder.pop();